test: 🔄 move testMarkRequired to a separate file
